pyproject.toml
[tool.poetry]
name = "asrom-llm"
version = "0.1.0"
description = ""
authors = ["Zijian Wu <zijian.wu14@gmail.com>"]
readme = "README.md"
packages = [{include = "asrom_llm"}]

[tool.poetry.dependencies]
python = "^3.10"
openai = "^0.27.2"
langchain = "^0.0.115"
pandas = "^1.5.3"
biopython = "^1.81"
tiktoken = "^0.3.2"
transformers = "^4.27.1"
sentence-transformers = "^2.2.2"
pubmed-parser = {git = "https://github.com/titipata/pubmed_parser.git"}
pymilvus = "2.2.0"




[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.0.0"
pre-commit = "^3.1.1"
flake8 = "^6.0.0"
black = "^23.1.0"
isort = "^5.12.0"
ipykernel = "^6.21.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 79
target-version = ['py38']
include = '\.pyi?$'
exclude = '''

(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | foo.py           # also separately exclude a file named foo.py in
                     # the root of the project
)
'''

[tool.isort]
profile = "black"
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
line_length = 99
